+++ chia/wallet/wallet_rpc_api.py
@@ -1367,8 +1367,8 @@ class WalletRpcApi:
 
         transactions = await self.service.wallet_state_manager.tx_store.get_transactions_between(
             wallet_id=request.wallet_id,
-            start=uint16(0) if request.start is None else request.start,
-            end=uint16(50) if request.end is None else request.end,
+            start=uint32(0) if request.start is None else request.start,
+            end=uint32(50) if request.end is None else request.end,
             sort_key=request.sort_key,
             reverse=request.reverse,
             to_puzzle_hash=to_puzzle_hash,
@@ -2162,7 +2162,7 @@ class WalletRpcApi:
         ) = await self.service.wallet_state_manager.trade_manager.trade_store.get_trades_count()
 
         return GetOffersCountResponse(
-            total=uint16(total), my_offers_count=uint16(my_offers_count), taken_offers_count=uint16(taken_offers_count)
+            total=uint32(total), my_offers_count=uint32(my_offers_count), taken_offers_count=uint32(taken_offers_count)
         )
 
     @tx_endpoint(push=True)
@@ -2618,7 +2618,7 @@ class WalletRpcApi:
                 transactions=[],
                 wallet_id=list(nft_dict.keys()),
                 spend_bundle=WalletSpendBundle([], G2Element()),
-                tx_num=uint16(len(interface.side_effects.transactions)),
+                tx_num=uint32(len(interface.side_effects.transactions)),
             )
 
     @tx_endpoint(push=True)
@@ -2683,7 +2683,7 @@ class WalletRpcApi:
                 transactions=[],
                 wallet_id=list(nft_dict.keys()),
                 spend_bundle=WalletSpendBundle([], G2Element()),
-                tx_num=uint16(len(interface.side_effects.transactions)),
+                tx_num=uint32(len(interface.side_effects.transactions)),
             )
 
     @marshal
